// GestionStockFront-main/src/app/models/produit.model.ts
export class Produit {
    id?: any;
    nom_produit?: string;
    categorie?: any;
    quantite_stock?: number;
    mode?: string;
    quantite_entrer?: number;
    quantite_sortie?: number;
    sortie_par?: any;
    date_entrer_sortie?: string;
    alerte_quantite?: number;
    date_ajoute?: string;
}


export class StockCrud {
    id?: any;
    nom_produit?: any;
    categorie?: any;
    quantite_stock?: any;
    mode?: any;
    quantite_entrer?: any;
    quantite_sortie?: any;
    sortie_par?: any;
    date_entrer_sortie?: any;
    alerte_quantite?: any;
    date_ajoute?: any;
  }
  
  